main folder renamed to main-page
